Backend/classmeet1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
pyjwt==2.9.0
passlib==1.7.4
bcrypt==4.1.3
pandas==2.2.2
openpyxl==3.1.5
pydantic[email]==2.9.2
pytest